# Rename `vec_duplicate_id()`, and new `vec_duplicate_loc()`