--- a/man/vec_duplicate.Rd
+++ b/man/vec_duplicate.Rd
@@ -4,6 +4,7 @@
 \alias{vec_duplicate}
 \alias{vec_duplicate_any}
 \alias{vec_duplicate_detect}
+\alias{vec_duplicate_flg}
 \alias{vec_duplicate_loc}
 \alias{vec_duplicate_all}
 \title{Find duplicated values}
@@ -12,18 +13,24 @@ vec_duplicate_any(x)
 
 vec_duplicate_detect(x)
 
+vec_duplicate_flg(x, first = FALSE)
+
 vec_duplicate_loc(x)
 
 vec_duplicate_all(x)
 }
 \arguments{
 \item{x}{A vector (including a data frame).}
+
+\item{first}{A logical indicating if the first occurrence of a value should
+be considered a duplicate.}
 }
 \value{
 \itemize{
 \item \code{vec_duplicate_any()}: a logical vector of length 1.
 \item \code{vec_duplicate_all()}: a logical vector of length 1.
 \item \code{vec_duplicate_detect()}: a logical vector the same length as \code{x}.
+\item \code{vec_duplicate_flg()}: a logical vector the same size as \code{x}.
 \item \code{vec_duplicate_loc()}: an integer vector of the locations of the
 duplicates of each value.
 }
@@ -37,6 +44,8 @@ similar to \code{\link[=anyDuplicated]{anyDuplicated()}}.
 element of the vector is duplicated elsewhere. Unlike \code{\link[=duplicated]{duplicated()}}, it
 reports all duplicated values, not just the second and subsequent
 repetitions.
+\item \code{vec_duplicate_flg()}: returns a vector of logical flags describing if each
+element of the vector is a duplicate.
 \item \code{vec_duplicate_loc()}: returns an integer vector giving the location of
 the duplicates of each value, i.e. the second and subsequent
 occurrences of that value.
@@ -59,6 +68,16 @@ vec_duplicate_all(c(1, 2))
 vec_duplicate_all(c(NA, NA))
 
 x <- c(10, 10, 20, 30, 30, 40)
+
+# The default of `vec_duplicate_flg()` is similar to `duplicated()`. Both
+# don't consider the first occurrence to be a duplicate.
+vec_duplicate_flg(x)
+duplicated(x)
+
+# Use `first` to control whether or not the first occurrence should be
+# considered a duplicate
+vec_duplicate_flg(x, first = TRUE)
+
 vec_duplicate_detect(x)
 # Note that `duplicated()` doesn't consider the first instance to
 # be a duplicate