--- a/man/vec_duplicate.Rd
+++ b/man/vec_duplicate.Rd
@@ -12,7 +12,7 @@ vec_duplicate_any(x)
 
 vec_duplicate_flg(x, first = FALSE)
 
-vec_duplicate_loc(x)
+vec_duplicate_loc(x, first = FALSE)
 
 vec_duplicate_all(x)
 }
@@ -39,8 +39,7 @@ similar to \code{\link[=anyDuplicated]{anyDuplicated()}}.
 \item \code{vec_duplicate_flg()}: returns a vector of logical flags describing if each
 element of the vector is a duplicate.
 \item \code{vec_duplicate_loc()}: returns an integer vector giving the location of
-the duplicates of each value, i.e. the second and subsequent
-occurrences of that value.
+the duplicates of each value.
 }
 }
 \section{Missing values}{
@@ -70,11 +69,13 @@ duplicated(x)
 # considered a duplicate
 vec_duplicate_flg(x, first = TRUE)
 
-# The behavior of `duplicated()` is more similar
-# to `vec_duplicate_loc()`
+# `vec_duplicate_loc()` returns the locations corresponding to
+# the flags that `vec_duplicate_flg()` returns
 vec_duplicate_loc(x)
+which(vec_duplicate_flg(x))
 
-# Which can be considered the complement of `vec_unique_loc()`
+# When `first = FALSE`, it can be considered the
+# complement of `vec_unique_loc()`
 vec_unique_loc(x)
 }
 \seealso{