… NAMESPACE
… NEWS.md
… man/vec_duplicate_detect.Rd
… man/vec_duplicate_id.Rd
… src/init.c
… tests/testthat/test-dictionary.R
… tests/testthat/test-lifecycle-deprecated.R