Closes #763 

If approved, this is what the changes suggested in #763 would look like.