@DavisVaughan 2020-01-17
Okay @lionel- we now have:

```
vec_duplicate_any(x)
vec_duplicate_all(x)
vec_duplicate_flg(x, first = FALSE)
vec_duplicate_loc(x, first = FALSE)
```

With a nice relationship of:

```
which(vec_duplicate_flg(x)) == vec_duplicate_loc(x)
which(vec_duplicate_flg(x, first = TRUE)) == vec_duplicate_loc(x, first = TRUE)
```

And soft deprecation of:

- `vec_duplicate_id()` -> `vec_first_loc()`
- `vec_duplicate_detect()` -> `vec_duplicate_flg(first = TRUE)`

`vec_duplicate_flg(first = TRUE)` is also a decent bit faster than `vec_duplicate_detect()` because I store and reuse the hashes rather than looking them up twice.

```r
library(nycflights13)
library(vctrs)

set.seed(123)

flights_with_repeats <- sample(nrow(flights), 1e6, TRUE)

# old

bench::mark(vec_duplicate_detect(flights_with_repeats), iterations = 100)
#> # A tibble: 1 x 6
#>   expression                                    min median `itr/sec` mem_alloc
#>   <bch:expr>                                 <bch:> <bch:>     <dbl> <bch:byt>
#> 1 vec_duplicate_detect(flights_with_repeats) 82.1ms 88.9ms      11.0    23.6MB
#> # … with 1 more variable: `gc/sec` <dbl>

# new

bench::mark(vec_duplicate_flg(flights_with_repeats, first = TRUE), iterations = 100)
#> # A tibble: 1 x 6
#>   expression                                               min median `itr/sec`
#>   <bch:expr>                                            <bch:> <bch:>     <dbl>
#> 1 vec_duplicate_flg(flights_with_repeats, first = TRUE) 39.6ms 42.1ms      23.2
#> # … with 2 more variables: mem_alloc <bch:byt>, `gc/sec` <dbl>
```

And if you don't want the first occurrence to be considered a duplicate, it is even faster:

```r
bench::mark(vec_duplicate_flg(flights_with_repeats, first = FALSE), iterations = 100)
#> # A tibble: 1 x 6
#>   expression                                                min median `itr/sec`
#>   <bch:expr>                                             <bch:> <bch:>     <dbl>
#> 1 vec_duplicate_flg(flights_with_repeats, first = FALSE) 25.6ms 30.6ms      32.4
#> # … with 2 more variables: mem_alloc <bch:byt>, `gc/sec` <dbl>
```