@hadley 2020-01-17
FWIW I really don't like the `flg` prefix