@DavisVaughan 2020-01-21
Postponing until 0.3.0 until we can figure out suffix naming for `flg`