@DavisVaughan 2020-03-18
It has come up elsewhere that we could also use the naming scheme:

- `vec_duplicate_detect` -> `vec_detect_duplicate()`

- `vec_duplicate_id()` -> `vec_locate_first()`

- `vec_duplicate_loc()` -> `vec_locate_duplicate()`

I think this also lends itself well to the proposed change of `vec_duplicate_any()` to `vec_any_duplicate()`, and the addition of `vec_all_duplicate()` and `vec_all_equal()`.

Also should think about maybe:

- `vec_unique_loc()` -> `vec_locate_unique()`

- `vec_unique_count()` -> `vec_count_unique()`

- `vec_group_id()` -> `vec_identity_group()`

- `vec_group_loc()` -> `vec_locate_group()`