@hadley 2020-03-18
`vec_locate_first()` doesn't really suggest first duplicate to me.  Do you mind making a googlesheet with a few of the proposals on it?